Added writing and reading of IDatetimeZones and hooked up the ZoneInfoCompiler to write the time zones it generates out. Currently the output format is as byte[] in resource files. If should be possible to change that if necessary but it make loading them from the assembly simpler. Currently only loading from the assembly is supported but will be adding support for loading from external files in the future.
At this point there is still much to do but I'm opening up everything I've done for
code reviewing if you want. If you see anything that is really ugly and without comments
then I probably havent' gotten there yet so be kind. I'm still in the process of
reformatting the code to match the "will of the people" but I may have missed a few files.