Store the min and max offsets as ticks, and add them manually.
We internally store Offset values as milliseconds so that we can use an int
field instead of a long; this means we need to perform a multiplication every
time we add one, which we do on every GetZoneIntervals call. We may wish to
investigate the benefit of introducing a TickOffset type for places where
we're going to use the same offset regularly in this way, only for classes
where we're not creating new instances very often (e.g. ZoneRecurrence and
*possibly* ZoneInterval).