Fix for issue 28 - handle fixed negative offsets.
ZoneTransition previously adjusted its offset to make sure that computation of
the local instant at the start of the transition would succeed; that then left
the offset incorrect for other times. We now keep the offset without any
checking, but provide a LocalInstant property which rounds on overflow.

All of this requires a new time zone database, as the error occurred while
creating it.