Fix issue 74: document nullability (and null-ness) consistently.
Remove any 'must not be null'-style text in parameter descriptions, replacing
it with exception documentation for ArgumentNullException instead.
(There may be cases where ArgumentNullException is not documented, and should
be, but this is enough to close this issue.)

Also consistently use the plaintext 'null' in documentation rather than
<c>null</c> or <see langword="null"/>, and consistently use '[param] is null.'
as the description for ArgumentNullException.

Finally, remove bogus documentation for a few Equals() methods that documented
that NullReferenceException would be thrown if called with null (it shouldn't,
and wouldn't - except in the case of ZoneAlias, where we fix that as a bug).