Remove the concept of missing vs zero values from Period. Fixes issue 90.
This removes the Units property entirely, makes the Year (etc) properties
on PeriodBuilder non-nullable, and changes period formatting to unconditionally
exclude zero-valued properties.

It also changes the internal representation of a Period from (units, long value,
long[] values) to individual (year, month, weeks, ...) fields.

The optimisation of using a single value in some cases seems unlikely to help -
we believe that most Period instances will be composite - and while the generic
long[] representation is helpful (and is still used temporarily in a few
places), the more explicit layout leads to less indirection in many cases (at
some cost to verbosity, it's true).

Also:
- Renamed Period.Empty to Period.Zero.
- Documented that Period.Normalize() can throw OverflowException.
  (Fixes issue 92)