Added a "Signed Release" configuration with strong naming. This relies on a file called "NodaTime Release.snk" in the "root" directory; this file is not present in source control, although the public key is.
The NuGet build now tries to build the Signed Release configuration, and
packages its binaries.

As the CommandLine library isn't signed, I've moved to including the source
directly in ZoneCompiler. If we ever ship the binaries of ZoneCompiler, we
probably need to ship the licence for it at the same time.