Simplify the (NodaZoneDate) serialised form of ZoneYearOffset to encode (possibly-negative) day-of-month using zigzag encoding.
This is somewhat unnecessary at present, but avoids the need to document
a bias value (and leads to a somewhat simpler format to document,
especially if we start to use this anywhere else).

This does not affect the size of the NodaZoneData file at all, though it
does make it marginally easier to interpret by hand.