Add a header and footer containing next/previous links to the generated user and developer guides, along with an auto-generated TOC and version number.
This changes some titles to be a bit more consistent between TOC and section
title, but otherwise is essentially a no-op as far as content is concerned.

Note that this doesn't add an 'Up' link to each page, but after playing
about for a while, I'm not sure that's worthwhile.

Fixes issue 127.