Added Trusted annotations where suggested by tooling - it all makes sense as far as I can see. (More to come potentially, as tooling improves.)
Duration deserialization has changed to be more consistent with Trusted values
- we were validating a value *after* using it in a trusted context; it would
have been okay in this case, but it's best to be consistent. The argument thrown
if the serialized data is invalid has been changed to plain ArgumentException,
as this is probably more sensible from the perspective of a caller.

One parameter is now *not* explicitly trusted, as it's not clear it makes
much sense in this context. A LocalDateTime value should always be valid in its
own right. This highlights the need for more clarity about exactly what Trusted
means.