New tool to dump time zones in tzvalidate format.
This may eventually be rolled into TzdbCompiler, although notably we don't need to provide a CLDR file to generate the tzvalidate format - and we can do it straight from an nzd file too, so it may be worth keeping them separate.

Usefully though, the tool can process a .tar.gz file (fetching it from the web if appropriate) which may be functionality worth including in TzdbCompiler.