Update TzdbCompiler to be flexible in its data source.
It can now load from a local directory, or a local archive, or a remote (ftp/http/https) archive.
This means NodaDump and the build scripts can be simplified too.

In a later commit I'll probably remove the tzdb data directory, as it no longer serves much purpose.

This change requires TzdbCompiler to target .NET 4.5, to avoid conflicts with LinqBridge in SharpCompress.