Stop DaylightSavingsDateTimeZone from being a DateTimeZone.
This means we no longer have to pass in bogus IDs all the time which make no sense.
Admittedly we need something *just* more than IZoneIntervalMap, so I've introduced IZoneIntervalMinMaxOffset which just adds the Min/MaxOffset properties, and may DateTimeZone implement that.
After this commit, we can rename the type to something saner...