Test more calculators in YearMonthCalculatorTest... and fix them when they break.
HebrewYearMonthCalculator was broken for year 0 and earlier - which isn't a problem for
most code, but it's possible that in a few cases our internals need to be able to handle values
outside the normal range. The cache format was assuming it could tweak the two high bits
arbitrarily - which doesn't work so well when the original value is negative. We now use the
bottom two bits for the lengths of Heshvan/Kislev.