"Partial patterns" for local date/time/date-time in ZonedDateTime and OffsetDateTime.
Allow the local date, local time, or local date/time parts of ZonedDateTimePattern and OffsetDateTimePattern to be specified as whole patterns, so they can use standard patterns.

It would be logical to allow this within LocalDateTime as well, but it would rarely be as useful, and then we'd need to worry about nested patterns.