More Unix epoch member consistency.
- Make the Instant.Ticks property a method (ToUnixTimeTicks)
- Rename FromTicksSinceUnixEpoch to FromUnixTimeTicks

This commit is a pure refactoring.

This will be a big breaking change to some clients, but it's useful to indicate that Ticks is no longer a sort of "inherent truth" property. (The fact that we're internally Unix epoch based is a little more encapsulated, effectively.)
We probably want to do the same thing with Duration.Ticks at some point, as it's quite different to the other Duration properties. Will file an issue.

Documentation changes coming in next commit.