Fix our Travis build, by skipping tests failing on Mono
Basically Mono doesn't play well with TimeZoneInfo in general. The NUnitLite
test runner fails if there's a test case source with no tests, so here we insert
a dummy one and explicitly use Assert.Ignore so we know the test is ignored.
Core CLR adopting IANA on Linux makes everything that bit trickier - this is
already covered in issue #434.