Fix the building of API docs.
This is still a set of interim measures until we move to docfx, but it's a start.
This includes removing tools projects which we don't need any more. SetVersion
isn't needed in exactly its current form, but we'll probably want to have something similar,
so I've kept it for now.