Convert WeekYearRule into an interface
This introduces WeekYearRuleExtensions (in the same source file as IWeekYearRule) to pass in CalendarSystem.Iso automatically.
(Other options would be to retain WeekYearRule as an abstract class that implements the interface, where the interface would
have two extra methods, or to use optional parameters.)

The static Iso property and factory method have moved to RegularWeekYearRule.

Some documentation changes, but incomplete.