Add dependency required for BigInteger to NodaTime.Testing.
I don't *think* this should be required - it should pick it up
automatically. There's no problem on Windows, but in .NET Core on Linux
it fails (in a hard-to-diagnose way when running tests - the tests never
fully start, with high CPU but no error). We should revisit this after
the next update.