Implement BCL week rules.
It turns out these are pretty simple:
- Either 1, 4 or 7 as the "min days in first week"
- Variable "first day of week"
- Unlike the ISO rules, the end of the calendar year is always in the same
  week-year, so the first week of the week-year may be shorter than 7 days.
  (The first days of the calendar year *may* be in the previous week-year
  though.)

There's a lot of common code between this and RegularWeekYearRule; it shouldn't
be too hard to refactor the code into a single class, providing the ability
for "regular" week year rules to have a variable first day of week too. That
can come later though, along with a whole user guide page with many, many
examples.