Move to use ASP.NET Core.
The benchmarks code has gone, as the format we were using is no longer
what we *will* be using, and there's no replacement just yet. The
benchmark data directory in the nodatime.org repo can die.

This is mostly just the template ASP.NET MVC web app, but with the
home controller removed (as we'll be serving static content) and a
new TzValidate controller which will start doing useful work when
the basics of "get the web site working" are handled.