Implement additional equality comparers for DateInterval.
Specifically, add DateInterval.NormalizingEqualityComparer and
ContainedDatesEqualityComparer.

NormalizingEqualityComparer additionally considers as equal two
intervals that differ only in whether they are closed or half-open
(inclusive or exclusive), meaning that [2001-01-01, 2001-01-31] is
considered equal to [2001-01-01, 2001-02-01).  However, empty intervals
with different start dates are still considered unequal.

(This is documented in terms of normalizing all intervals to exclusive
intervals, since that's mathematically simpler; however, in practice we
cannot actually represent all inclusive intervals as exclusive, and so
we implement this by normalizing in the other direction, then treating
empty intervals as a special case.)

ContainedDatesEqualityComparer additionally considers all empty
intervals (for the same calendar) to be equal, effectively implementing
equality on the set of dates contained within the interval.

Fixes #383.