Increase the number of bits available to months
We now support months in the range [1, 32]. This is
necessary for the Wondrous calendar which has 19 months.

I've taken the extra bit from the calendar range, so we can now "only"
support 64 calendars... but we've only got 18 at the moment, and the
Wondrous calendar will be the 19th. I think that's safe.

This needs to be done before 2.0 as it affects the serialization
format.