Re-enable all time zones for Mono tests
There are 14 broken tests on Mono for Linux; skipping
category BrokenOnMonoLinux will skip these. (They are
all due to bugs in Mono, as far as we can tell.) This
also skips tests for time zones which use Feb 29th in
any TimeZoneInfo rules due to another bug in Mono.

This helps narrow down #97.