Emergency changes to Noda Time web site
Once we've got Google Cloud Storage working from Azure again, all
we need to do is change the comments in Startup.cs - we should keep
the fakes where they are for future use.

Note that when building the web site at the moment I'm also copying
the nzd files in manually. That will just stop when it's fixed.