Use an environment variable to determine whether to use GCS.
By default (as per now), we'll use Google Cloud Storage for listing and
fetching the downloads. However, if DISABLE_GCS is set (to any value),
we'll avoid initialising GCS and instead rely on the fake repositories.

This allows one to run:

$ DISABLE_GCS=1 dotnet run

without any additional code changes, and get a working local site.

Also avoid registration of the GoogleCredential singleton if GCS is
disabled, since it's only used by the GCS-backed repositories, and has a
non-trivial setup.