Fix Period computation bug
The approach taken before was fundamentally unsound. It's clearer to
consider that when we get down to time-based fields, they're all
fixed-length, so we only need to consider the duration between the
start and end points. This is an initial fix to patch into 2.0.2 -
it's not as efficient as it might be, but it's reasonably minimal.

Further refactoring will come in a separate PR, but that can stay on
the master branch only.

Fixes #824.