Added IWeekYearRule to 1.4.x.
This will allow us to make GetGregorian(int) etc obsolete while
providing 2.0.x-compatible alternatives.

The implementation and tests are as close to 2.0.x as I could make
them - they're not necessarily how I'd normally write them with 1.x,
but I wanted to make as few changes as possible. (Hence the multiple
conversions between ticks and days, because 2.0 is day-based.)