Improve error message for invalid standard patterns.
Sample:

Unhandled Exception: NodaTime.Text.InvalidPatternException:
The standard format "H" is not valid for the NodaTime.LocalDateTime
type. If the pattern was intended to be a custom format, escape it
with a percent sign: "%H".

I haven't implemented the "detect whether it's a valid custom
pattern" part - I think that would be a bigger (and slightly
riskier) change without much benefit.

Fixes #971.