Fetch docfx version explicitly
The change to the annotation generator is due to a new version of
docfx which expects a comment at the start of each yml file (but the
comment isn't preserved by a load/save cycle).

I've got the history branch ready to push just before I merge this PR.

Fixes #938.