Update TZDB downloads to https
I have a slight concern that this may trigger clients to get
confused, thinking that something's changed but then being surprised
when nothing has.

One option would be to use http for anything at or prior to 2017c,
but https from 2017d onwards.

Fixes #999. (I believe this is all there is left.)