Fix issues when TimeZoneInfo.Local is null
This was found when trying out C# 8 nullable references.
TimeZoneInfo.Local has been observed to return null in some Mono
installations. We should handle that as reasonably as we can. We
already did for BclDateTimeZoneSource, but now we do in
TzdbDateTimeZoneSource too. BclDateTimeZone.ForSystemDefault needed
fixing too.

Added tests for everywhere we need to
use TimeZoneInfo.Local (via the interceptor).