Get the snippet extractor "mostly" working again
Life is made harder by NodaTime.dll and NodaTime.Demo.dll no longer targeting the desktop framework, while the snippet extractor has to due to its MSBuild dependencies.
The workaround is a bit of a hack, and we lose the output - but at least we get the snippets themselves.
Will talk to the Roslyn team about what to do with this...