Optimize OffsetTime representation
OffsetDateTime currently uses a single 64-bit integer field for
offset and nano-of-day; we can do the same thing in OffsetTime.
Once this change is in, we can experiment with changing
OffsetDateTime to use OffsetTime instead of its own field, and maybe
use LocalDate instead of its own YearMonthDayCalendar, due to
improvements with readonly structs.