Add type converters for ISO pattern parsers
Npgsql uses type converters for bound-parsing in `NpgsqlRange<T>`
which represents a PostgreSQL range type.

The goal is to add type converters where they make sense, such
as for types that have ISO patterns that are (reasonably) safe to
use for roundtrip serialization.

No effort is made to handle non-ISO calendars on write, because
there would be no way to detect and handle them on read.