Add type converter unit tests
This also changes the constructors from private to public and the
`TypeConverterAttribute` declarations were updated to use `typeof`
rather than `nameof`.

When `typeof` is used, the `TypeConverterAttribute` stores the
assembly qualified name, while `nameof` just stores the string
directly (in this case, without qualification).

Without the qualification, the `TypeDescriptor` fails to locate
the derived converter and falls back to the base implementation.
This caused earlier ad-hoc testing to pass, because the the base
implementation was returned (and thus not null).

The unit tests commited here are designed to fail if the base
implementation is returned by `TypeDescriptor.GetConverter(...)`.