Add Duration.ToInt128Nanoseconds and Duration.FromNanoseconds(Int128)
These are only available for the .NET 8 build (although in source code they're .NET 7+ as that's when Int128 was introduced; this is just in case we introduce a .NET 7 build).

Fixes #1777.

Benchmarks show that the Int128 operations are about the same as BigInteger in the worst cases, and about twice as fast in better cases.