Use invariant formatting for all interpolated strings
Where culture-specific formatting is required, it's done internally (i.e. the values which are interpolated are preformatted).

In some cases this may have a small performance penalty, but most of the time this is when an exception is being thrown anyway.

We should check benchmarks for DateInterval.ToString and Interval.ToString - simple string concatenation may be worthwhile.